Rename twitch apiwrappers
